requirements.txt
asn1crypto
cffi
cryptography
iniherit
psycopg2
pycparser
PyMySQL
six
text-unidecode
yoyo-migrations